backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.0
requests>=2.31.0
python-dotenv>=1.0.0
spotipy>=2.23.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
